requirements.txt
Django==5.0.2
mysqlclient==2.2.4
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-tailwind==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
mysql-connector-python==8.3.0 